+++ packages/components/components/elvis-carousel/src/react/StyledComponents.tsx
@@ -1,10 +1,11 @@
 import styled from 'styled-components';
+import * as ElviaColors from '@elvia/elvis-colors';
 
-const ElviaColors = {
-  elviaCharge: '#29d305',
-  elviaOn: '#ffffff',
-  elviaOff: '#000000',
-  elviaBlue: '#0064fa',
+const colors = {
+  elviaCharge: ElviaColors.default['primary-colors']['green'].color,
+  elviaOn: ElviaColors.default['primary-colors']['white'].color,
+  elviaOff: ElviaColors.default['primary-colors']['black'].color,
+  elviaBlue: ElviaColors.default['internal-colors']['focus-outline'].color,
 };
 
 const mobileMax = '767px';
@@ -90,7 +91,7 @@ export const CarouselElement = styled.div<CarouselElementTypes>`
   font-family: 'Red Hat Text', Verdana, sans-serif;
   line-height: '28px';
   font-size: '16px';
-  color: ${ElviaColors.elviaOff};
+  color: ${colors.elviaOff};
   text-transform: 'unset';
   letter-spacing: 'unset';
   padding-bottom: '10px';
@@ -100,7 +101,7 @@ export const CarouselElement = styled.div<CarouselElementTypes>`
     font-family: 'Red Hat Text', Verdana, sans-serif;
     line-height: '28px';
     font-size: '16px';
-    color: ${ElviaColors.elviaOff};
+    color: ${colors.elviaOff};
     text-transform: 'unset';
     letter-spacing: 'unset';
     font-style: unset;
@@ -134,25 +135,24 @@ export const ListOfDots = styled.div`
 
 export const Dot = styled.button`
   border: 1px solid
-    ${(props: { isSelected: boolean }) => (props.isSelected ? ElviaColors.elviaCharge : ElviaColors.elviaOff)};
+    ${(props: { isSelected: boolean }) => (props.isSelected ? colors.elviaCharge : colors.elviaOff)};
   height: ${(props: { isSelected: boolean }) => (props.isSelected ? '9px' : '8px')};
   width: ${(props: { isSelected: boolean }) => (props.isSelected ? '9px' : '8px')};
   @media (max-width: 767px) {
     border: 1px solid
-      ${(props: { isSelected: boolean }) =>
-        props.isSelected ? ElviaColors.elviaCharge : ElviaColors.elviaOff};
+      ${(props: { isSelected: boolean }) => (props.isSelected ? colors.elviaCharge : colors.elviaOff)};
     height: ${(props: { isSelected: boolean }) => (props.isSelected ? '13px' : '12px')};
     width: ${(props: { isSelected: boolean }) => (props.isSelected ? '13px' : '12px')};
   }
 
   border-radius: 50%;
   background-color: ${(props: { isSelected: boolean }) =>
-    props.isSelected ? ElviaColors.elviaCharge : ElviaColors.elviaOn};
+    props.isSelected ? colors.elviaCharge : colors.elviaOn};
   margin: ${(props: { isSelected: boolean }) => (props.isSelected ? '7.5px' : '8px')};
   cursor: pointer;
   padding: 0;
   &:hover {
-    background-color: ${ElviaColors.elviaCharge};
+    background-color: ${colors.elviaCharge};
   }
 `;
 
@@ -177,7 +177,7 @@ export const LeftCarouselButton = styled.button`
     background-position: center;
     background-repeat: no-repeat;
     background-image: url("data:image/svg+xml,%3csvg viewBox='0 0 24 24' aria-hidden='true' width='24' height='24' fill='none' xmlns='http://www.w3.org/2000/svg'%3e%3cpath fill-rule='evenodd' clip-rule='evenodd' d='M6.4 12a.8.8 0 01.8-.8h9.6a.8.8 0 110 1.6H7.2a.8.8 0 01-.8-.8z' fill='black'/%3e%3cpath fill-rule='evenodd' clip-rule='evenodd' d='M11.766 7.434a.8.8 0 010 1.132L8.33 12l3.435 3.434a.8.8 0 11-1.132 1.132l-4-4a.8.8 0 010-1.132l4-4a.8.8 0 011.132 0z' fill='black'/%3e%3cpath fill-rule='evenodd' clip-rule='evenodd' d='M12 1.6C6.256 1.6 1.6 6.256 1.6 12c0 5.744 4.656 10.4 10.4 10.4 5.744 0 10.4-4.656 10.4-10.4 0-5.744-4.656-10.4-10.4-10.4zM0 12C0 5.373 5.373 0 12 0s12 5.373 12 12-5.373 12-12 12S0 18.627 0 12z' fill='%2329D305'/%3e%3c/svg%3e");
-    background-color: ${ElviaColors.elviaOn};
+    background-color: ${colors.elviaOn};
     display: inline-block;
     height: 32px;
     width: 32px;
@@ -204,7 +204,7 @@ export const RightCarouselButton = styled.button`
     background-position: center;
     background-repeat: no-repeat;
     background-image: url("data:image/svg+xml,%3csvg viewBox='0 0 24 24' aria-hidden='true' width='24' height='24' fill='none' xmlns='http://www.w3.org/2000/svg'%3e%3cpath fill-rule='evenodd' clip-rule='evenodd' d='M13.366 7.434a.8.8 0 10-1.132 1.132L14.87 11.2H7.2a.8.8 0 100 1.6h7.669l-2.635 2.634a.8.8 0 101.132 1.132l3.996-3.996a.77.77 0 00.166-.237.798.798 0 00-.166-.902l-3.996-3.997z' fill='black'/%3e%3cpath fill-rule='evenodd' clip-rule='evenodd' d='M12 22.4c5.744 0 10.4-4.656 10.4-10.4 0-5.744-4.656-10.4-10.4-10.4C6.256 1.6 1.6 6.256 1.6 12c0 5.744 4.656 10.4 10.4 10.4zM24 12c0 6.627-5.373 12-12 12S0 18.627 0 12 5.373 0 12 0s12 5.373 12 12z' fill='%2329D305'/%3e%3c/svg%3e");
-    background-color: ${ElviaColors.elviaOn};
+    background-color: ${colors.elviaOn};
     display: inline-block;
     height: 32px;
     width: 32px;
@@ -220,7 +220,7 @@ export const CheckButton = styled.button`
   cursor: pointer;
   &:hover {
     i {
-      background-color: ${ElviaColors.elviaCharge};
+      background-color: ${colors.elviaCharge};
     }
   }
   i {