+++ packages/components/components/elvis-box/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Elvia Box Changelog
 
+## 1.1.4 (19.11.21)
+
+### Bug fix
+
+- Now using elvis-colors pack for colors.
+
 ## 1.1.0 (03.11.21)
 
 ### Feature